Marker now requires token for every mutation

+++ src/api/resolvers/markerResolver.ts
@@ -1,6 +1,8 @@
+import CustomError from '../../classes/CustomError';
 import {Marker} from '../../interfaces/Marker';
 import {UserIdWithToken} from '../../interfaces/User';
 import markerModel from '../models/markerModel';
+import {GraphQLError} from 'graphql';
 
 export default {
   Query: {
@@ -23,22 +25,40 @@ export default {
         ' and user ',
         user
       );
-      /* if (!user.id) {
+      if (!user.id) {
         throw new GraphQLError('Not authorized', {
           extensions: {code: 'NOT_AUTHORIZED'},
         });
       }
-      args.owner = user.id; */
+      args.owner = user.id;
       const marker = new markerModel(args);
       return await marker.save();
     },
-    updateMarker: async (_parent: undefined, args: Marker) => {
-      return await markerModel.findByIdAndUpdate(args.id, args, {
-        new: true,
-      });
+    updateMarker: async (
+      _parent: undefined,
+      args: {marker: Marker; token: string}
+    ) => {
+      if (args.token === 'greatest-secret-token') {
+        return await markerModel.findByIdAndUpdate(
+          args.marker.id,
+          args.marker,
+          {
+            new: true,
+          }
+        );
+      } else {
+        return new CustomError('Forbidden', 401);
+      }
     },
-    deleteMarker: async (_parent: undefined, args: {id: string}) => {
-      return await markerModel.findByIdAndDelete(args.id);
+    deleteMarker: async (
+      _parent: undefined,
+      args: {id: string; token: string}
+    ) => {
+      if (args.token === 'greatest-secret-token') {
+        return await markerModel.findByIdAndDelete(args.id);
+      } else {
+        return new CustomError('Forbidden', 401);
+      }
     },
   },
 };
